requirements.txt
argh
ccxt
configobj
invoke
marrow.mailer
meld3
objectpath
pyDAL
pyrogram>=0.7.4
python-dateutil
retry
supycache
